Patched a nasty bug (low-res image generator bug)
This bug would render the resized images useless to upscale back, as quality would be lost.
The patch relies on ImageDrawable holding a copy of the original image, that it then uses to resize.
What needs to be explored is the impact of this change to memory usage and (well), performance in general.